feat: - modularize scripts - reload page on filter change - url changes to shortlink - content creator panel changes
fix: cleanup, bad code

fix: restore reverted code
